pyproject.toml
[project]
name = "youtube-bulkup"
version = "0.1.0"
description = "Efficient YouTube Bulk Uploader with Automated Metadata Extraction"
authors = [
    {name = "Antigravity", email = "antigravity@google.com"},
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "google-api-python-client>=2.111.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth>=2.26.0",
    "tenacity>=8.2.3",
    "xxhash>=3.4.1",
    "tinydb>=4.8.0",
    "python-dotenv>=1.0.0",

    "hachoir>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning:httplib2.*",
    "ignore::DeprecationWarning:pkg_resources.*",
    "ignore::ResourceWarning:httplib2.*",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]


[project.scripts]
yt-up = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
